Extract CommentOptions, rename fields
