+++ src/main/kotlin/io/github/mojira/arisa/domain/CommentOptions.kt
@@ -1,0 +1,7 @@
+package io.github.mojira.arisa.domain
+
+data class CommentOptions(
+    val messageKey: String,
+    val variable: String? = null,
+    val language: String = "en"
+)
