+++ src/main/kotlin/io/github/mojira/arisa/infrastructure/jira/Functions.kt
@@ -25,12 +25,6 @@ fun getIssue(jiraClient: JiraClient, key: String) = runBlocking {
 }
 const val MILLI_FOR_FORMAT = 123L
 
-data class CommentOptions(
-    val key: String,
-    val filledText: String? = null,
-    val lang: String = "en"
-)
-
 fun updateCHK(issue: Issue, chkField: String): Either<Throwable, Unit> = runBlocking {
     Either.catch {
         issue
