+++ src/main/kotlin/io/github/mojira/arisa/modules/CrashModule.kt
@@ -7,7 +7,7 @@ import arrow.core.firstOrNone
 import io.github.mojira.arisa.domain.Attachment
 import io.github.mojira.arisa.domain.Issue
 import io.github.mojira.arisa.infrastructure.config.CrashDupeConfig
-import io.github.mojira.arisa.infrastructure.jira.CommentOptions
+import io.github.mojira.arisa.domain.CommentOptions
 import me.urielsalis.mccrashlib.Crash
 import me.urielsalis.mccrashlib.CrashReader
 import me.urielsalis.mccrashlib.parser.ParserError
