+++ src/main/kotlin/io/github/mojira/arisa/modules/KeepPrivateModule.kt
@@ -7,7 +7,7 @@ import arrow.core.right
 import io.github.mojira.arisa.domain.ChangeLogItem
 import io.github.mojira.arisa.domain.Comment
 import io.github.mojira.arisa.domain.Issue
-import io.github.mojira.arisa.infrastructure.jira.CommentOptions
+import io.github.mojira.arisa.domain.CommentOptions
 import java.time.Instant
 
 class KeepPrivateModule(
